/srv/dev-disk-by-label-120GB/torrent/Server added

+++ scripts/torrent/mp3/removeArtists2.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 while IFS='' read -r line || [[ -n "$line" ]]; do
-    find . -xtype f -iname *"$line"* -print -delete
-    #echo "$line"
+    if [[ "$line" == *" "* ]]; then
+        find . -xtype f -iname *"$line"* -print -delete
+    fi
 done < "$1"
